regime-monitor/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
asyncpg==0.30.0
python-dotenv==1.0.1
websockets==14.0
